requirements.txt
# Core LLM dependencies
openai>=1.0.0
anthropic>=0.8.0

# CRM integrations
simple-salesforce>=1.12.0
hubspot-api-client>=8.0.0

# Email and outreach
sendgrid>=6.10.0
python-dotenv>=1.0.0

# Web scraping and research
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data handling
pydantic>=2.0.0
python-dateutil>=2.8.0

# Logging and monitoring
structlog>=23.0.0
